+++ src/elements/author/src/index.tsx
@@ -36,7 +36,7 @@ const Author: React.FC<Props> = ({
         display: 'flex',
         alignItems: 'center',
         flexDirection: 'row',
-        variant: 'author.main'
+        variant: 'elements.author.main'
       }}
       className={className}
     >
@@ -48,7 +48,7 @@ const Author: React.FC<Props> = ({
             height: [48, 56],
             width: [48, 56],
             marginRight: 'sm',
-            variant: 'author.image-link'
+            variant: 'elements.author.image-link'
           }}
         >
           <authorImage.type
@@ -57,7 +57,7 @@ const Author: React.FC<Props> = ({
               maxHeight: '100%',
               maxWidth: '100%',
               borderRadius: '50%',
-              variant: 'author.image'
+              variant: 'elements.author.image'
             }}
           />
         </Styled.a>
@@ -70,7 +70,7 @@ const Author: React.FC<Props> = ({
             marginBottom: 'xxs',
             fontWeight: 'bold',
             paddingBottom: 0,
-            variant: 'author.heading'
+            variant: 'elements.author.heading'
           }}
           as="h2"
         >
@@ -79,24 +79,26 @@ const Author: React.FC<Props> = ({
             sx={{
               color: 'inherit',
               textDecoration: 'none',
-              variant: 'author.link'
+              borderBottom: 0,
+              variant: 'elements.author.link'
             }}
           >
-            {authorString}
+            Written by {authorString}
           </a>
         </Styled.h6>
         <Styled.p
           sx={{
             fontSize: 'xxs',
             marginY: 0,
-            variant: 'author.text'
+            variant: 'elements.author.text'
           }}
         >
           <Styled.a
             href={editorUrl}
             sx={{
               textDecoration: 'none',
-              variant: 'author.link'
+              borderBottom: 0,
+              variant: 'elements.author.link'
             }}
           >
             {editorName &&