+++ src/compounds/sponsored-product/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@1.0.3...@uswitch/trustyle.sponsored-product@2.0.0) (2020-07-17)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product
+
+
+
+
+
 ## [1.0.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@1.0.2...@uswitch/trustyle.sponsored-product@1.0.3) (2020-07-15)
 
 **Note:** Version bump only for package @uswitch/trustyle.sponsored-product