+++ src/compounds/tabs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.tabs@0.2.3...@uswitch/trustyle.tabs@1.0.0) (2020-07-17)
+
+**Note:** Version bump only for package @uswitch/trustyle.tabs
+
+
+
+
+
 ## [0.2.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.tabs@0.2.2...@uswitch/trustyle.tabs@0.2.3) (2020-07-13)
 
 **Note:** Version bump only for package @uswitch/trustyle.tabs